Also use end pieces in zig-zag infill pattern for single layer
In multi-layer infill, it added these end pieces properly, but not in single-layer infill. They should be added, because not adding them leaves loose endpoints for the infill which are a weak point. This is especially visible in 100% density since it leaves gaps then if the infill is slightly too thin to add another line.

I believe this was correct before, so I think a recent refactor accidentally set this to False.

Contributes to issue CURA-9077.